requirements.txt
langchain
langchain-ollama
langgraph
presidio-analyzer
presidio-anonymizer
spacy
sentence-transformers
numpy
scikit-learn
streamlit
pandas
plotly
altair
fastapi
uvicorn
pydantic
requests
python-dotenv
